fix: rolesUtils and useCanI
